+++ src/action/authorize/award/point.ts
@@ -3,18 +3,11 @@ import { ActionType } from '../../../actionType';
 import { TransactionType } from '../../../transactionType';
 import * as AuthorizeActionFactory from '../../authorize';
 
-// 最適化(2022-06-01~)
 export type IAgent = ActionFactory.IParticipantAsSeller;
-// 最適化(2022-06-01~)
 export type IRecipient = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
-
 export enum ObjectType {
     PointAward = 'PointAward'
 }
-
-/**
- * オーソリ対象
- */
 export interface IObject {
     typeOf: ObjectType;
     /**
@@ -30,22 +23,15 @@ export interface IObject {
      */
     notes?: string;
 }
-
 export interface IResult {
-    price: number;
+    // price: number; // discontinue(2023-11-27~)
     amount: number;
 }
-
 export interface IPurpose {
     typeOf: TransactionType;
     id: string;
 }
-
 export type IError = any;
-
-/**
- * ポイントインセンティブ承認アクション属性
- */
 export interface IAttributes extends AuthorizeActionFactory.IAttributes<IObject, IResult> {
     typeOf: ActionType.AuthorizeAction;
     object: IObject;
@@ -53,7 +39,6 @@ export interface IAttributes extends AuthorizeActionFactory.IAttributes<IObject,
     recipient: IRecipient;
     purpose: IPurpose;
 }
-
 /**
  * ポイントインセンティブ承認アクション
  * 注文取引のインセンティブとしてポイントを付与する場合に使用されます。