+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - update @waiter/factory
 - 興行オファー承認をオファー未指定に対応
+- 単価オファー検索条件拡張
 
 ### Deprecated